+++ NEWS.md
@@ -6,6 +6,7 @@
 
 ### Functions
 
+- `inst/extdata/co_unidade_ensino/co_unidade_ensino.csv` now has the code of all Brazilian federal institutes. (#107)
 - Fixed a bug in `read_sistec()` encode for linux. (#105)
 - Implemented `read_suap()`. (#99)
 - Implemented `read_conecta()`. (#97)